refactored graphql/route.ts
