+++ package.json
@@ -28,6 +28,7 @@
         "@aws-sdk/s3-request-presigner": "^3.645.0",
         "@googlemaps/js-api-loader": "^1.16.6",
         "@googlemaps/routing": "^1.1.0",
+        "@graphql-tools/schema": "^10.0.7",
         "@vis.gl/react-google-maps": "^1.1.0",
         "child_process": "^1.0.2",
         "dayjs": "^1.11.10",
