+++ src/components/shared/Avatar/helpers/getContent.tsx
@@ -1,4 +1,4 @@
-import { IUser } from "@/database/models/User";
+import { IUser } from "@/database/types/User";
 
 export const getContent = (name: string): string | null => {
     if (name[0] && name[1]) return `${name[0]}${name[1]}`.toUpperCase();
