+++ src/database/types/User.ts
@@ -1,0 +1,20 @@
+import { User } from "next-auth";
+import { IRole } from "./Role";
+
+export interface IUser {
+    _id: string;
+    name: string;
+    firstName: string;
+    lastName: string;
+    email: string;
+    image: string;
+    location: string;
+    aboutMe: string;
+    interests: string[];
+    meetings: string[];
+    tags: string[];
+    emailVerified: boolean;
+    createdAt: Date;
+    updatedAt: Date;
+    role: IRole;
+}
